fix: date filtering
