+++ multi-git-dashboard/src/components/cards/OverviewCard.tsx
@@ -1,6 +1,6 @@
+import { calculateCurrentWeek } from '@/pages/_app';
 import { RangeSlider, Stack } from '@mantine/core';
 import { TeamData } from '@shared/types/TeamData';
-import dayjs from 'dayjs';
 import { useState } from 'react';
 import Analytics from '../overview/analytics/Analytics';
 import PR from '../overview/pr/PR';
@@ -13,45 +13,28 @@ export interface OverviewProps {
   profileGetter: ProfileGetter;
 }
 
-export const START_DATE = dayjs('2024-01-15');
-export const BREAK_START_WEEK = 6;
-export const BREAK_DURATION_WEEKS = 1;
-export const weekToDates = (week: number) => {
-  let startDate = START_DATE.add(week, 'week');
-  if (week >= BREAK_START_WEEK) {
-    startDate = startDate.add(BREAK_DURATION_WEEKS, 'week');
-  }
-  return startDate;
-};
-
 export const OverviewCard: React.FC<OverviewProps> = ({
   team,
   teamData,
   teamDatas,
   profileGetter,
 }) => {
-  const calculateCurrentWeek = () => {
-    const currentWeek = dayjs().diff(START_DATE, 'week');
-    return currentWeek >= BREAK_START_WEEK
-      ? currentWeek + BREAK_DURATION_WEEKS
-      : currentWeek;
-  };
   const totalWeeks = calculateCurrentWeek();
 
   const [selectedWeekRange, setSelectedWeekRange] = useState<[number, number]>([
     0,
-    totalWeeks,
+    totalWeeks - 1,
   ]);
 
   return (
     <Stack>
       <RangeSlider
         value={selectedWeekRange}
-        max={totalWeeks}
+        max={totalWeeks - 1}
         minRange={1}
         onChange={setSelectedWeekRange}
         label={value => `Week ${value + 1}`}
-        marks={Array.from({ length: totalWeeks + 1 }, (_, i) => ({
+        marks={Array.from({ length: totalWeeks }, (_, i) => ({
           value: i,
           label: `Week ${i + 1}`,
         }))}
